XMPP Control Added
Added a primitive help message, about message, the compressor part, and the temerature controller part.
Finally fixed the issue with the compressor action parameter type.
It seems to be impossible to reply with several messages in a row. Not all the messengers support multiline messages. Shame on them.
Created config file to simplify search over the vast script.