Equipment doesn't work right. Diagnostic changes made.
The current source does not decreases it's voltage over time without a load connected. The load resistance should be some 1MOhm or less for the source to work fine.
The pressure gauge appears not to be connected sometimes although its display reads ‘PC’. No idea why and how to fix. Increased verbosity in the autostart tool.
The autostart request to the server to turn the compressor on returns Error 503 (Service Unavailable). No idea why either.