Minor improvements
Increased web interface update rate.
Increased thermal control communication rate from 1 second to 0.25 s. Will it hold?
Bug with favicon display fixed.
It's necessary to separate pages that control and/or monitor different devices. They should be accessible both at the dedicated web pages and at the main page within iframes.
It's necessary to save the thermal controller PID settings and restore them on the web page requested. The settings are sent to the server when a user enables a PID channel. It's considered that the user is pretty sure to apply the settings now and then.