Hide useless buttons and disable useless PID controls
If the compressor is running, Start button is disabled. If the compressor is not running, Stop button is disabled.
If the compressor fails, both Start and Stop buttons are disabled.
If a thermometer or an output of the thermal contoller fails, PID parameters and action buttons are hidden. This should be changed with options limited to what is present.