Improved UI
Added favicon.
Changed UI behaviour to disallow using PID only if there is neither a sensor nor an output present. It corresponds to what was marked to do last time.
It turns out that the temperature controller checks for an output to be plugged in only when it tries to use it.
A sound warning on client side is very desirable. So is an UBS for the server.