Commands are hidden into the backend. The temperature controller parameters get synchronized among clients.
The sequencies of commands issued to both the helium compressor and the temperature controller should not be visible to a client. Moreover, a client should not be able to inject and issue other commands.
As the first step to scheduled temperature run, the temperature controller PID parameters are now synchronized. The desired temperatures and the dedicated PID parameters will be synchronized, too. As for the data, the exact way to do it is not yet clear.