Minor UI fixes
Reduced plot height to fit small displays of 1024 × 768 px.
Iframes body margins are not set to 0, for it takes a lot of computations with no significant profit.
Temperature data update interval brought back to 1 second to decrease memory consumption and both server and client load. RTM could be of some use.
Fixed data type mismatch in the helium compressor request. To be considered again.