Numerous temperature controller UI changes
Added local sound alert when the temperature less than 77 K is reached and on warnings.
Fixed elements hiding rules to show only what is needed.
Removed unnecessary options in the PID settings.
Added manual PID output power setting.
Added error description shown in case of an error. The description is the same as written to the server log.
Fixed entering other settings issue when the PID is on.