Numerous changes
Added support for the newly arrived vacuum pumping station. The former file is kept but slightly renamed.
Added support for the relay module that can turn on both the water pump and the vacuum pump.
Fixed easter egg.
Minor interface fixes and changes.
Encountered a bizarre caching bug in Firefox. No idea how to avoid it in the future.
Except for the bug, everything is ready for the autostart feature.